fix: repair frontend crash after moving layout/ folder in pages/
