+++ frontend/src/organisms/Scenario/ScenarioList.tsx
@@ -3,7 +3,6 @@ import { useCurrentUser } from "@/lib/zustand/userStore";
 import type { Entities } from "@/types/entities";
 import { ScrollArea } from "@radix-ui/react-scroll-area";
 import { Link } from "react-router-dom";
-import ModalToAuth from "../../layout/ModalToAuth";
 import {
   Card,
   CardContent,
@@ -11,6 +10,7 @@ import {
   CardHeader,
   CardTitle,
 } from "../../lib/shadcn/generated/ui/card";
+import ModalToAuth from "../../pages/layout/ModalToAuth";
 import ModalToOpenScenario from "./ModalToOpenScenario";
 
 type Props = {
